test: improve coverage to keep tests passing
