+++ test/utils.ts
@@ -1,9 +1,12 @@
 import { strict as assert } from 'assert'
-import { createCanvas, createImageData } from 'canvas'
+import { createCanvas, createImageData, ImageData } from 'canvas'
 import { randomBytes } from 'crypto'
 import { promises as fs } from 'fs'
 import { Rect } from '../src/types'
-import { toRGBA } from '../src/utils/convert'
+import {
+  makeImageTransform,
+  assertGrayscaleImage,
+} from '../src/utils/imageFormatUtils'
 
 export function randomImage({
   width = 0,
@@ -69,6 +72,39 @@ export function randomInt(
 ): number {
   assert(min <= max)
   return (min + Math.random() * (max - min + 1)) | 0
+}
+
+/**
+ * Converts an image to an RGBA image.
+ */
+export const toRGBA = makeImageTransform(grayToRGBA, (imageData) =>
+  imageData instanceof ImageData
+    ? imageData
+    : createImageData(imageData.data, imageData.width, imageData.height)
+)
+
+/**
+ * Converts a grayscale image to an RGBA image.
+ */
+export function grayToRGBA(imageData: ImageData): ImageData {
+  assertGrayscaleImage(imageData)
+
+  const { data: src, width, height } = imageData
+  const dst = new Uint8ClampedArray(width * height * 4)
+
+  for (let y = 0; y < height; y += 1) {
+    for (let x = 0; x < width; x += 1) {
+      const srcOffset = x + y * width
+      const dstOffset = srcOffset << 2
+
+      dst[dstOffset] = src[srcOffset]
+      dst[dstOffset + 1] = src[srcOffset]
+      dst[dstOffset + 2] = src[srcOffset]
+      dst[dstOffset + 3] = 255
+    }
+  }
+
+  return createImageData(dst, width, height)
 }
 
 export async function writeImageToFile(
